refactor: Organizar vistas de matrícula en subdirectorio
Se han movido los archivos de vista de matrícula a un subdirectorio para mejorar la organización del proyecto.

- `views/matricula_nueva_view.php` ha sido movido a `views/matriculas/nueva.php`.
- `views/matriculas_view.php` (la lista principal) ha sido movido a `views/matriculas/list.php`.

Se ha actualizado el archivo `controllers/matriculas_controller.php` para que apunte a las nuevas ubicaciones de estos archivos, asegurando que la aplicación siga funcionando correctamente después de la refactorización.
